refactor : fix module structure for parser crate
