--- crates/parser/src/account_parser/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "token-program")]
-pub mod token_program;
-
-#[cfg(feature = "token-extensions")]
-pub mod token_extensions;
