--- crates/parser/src/account_parser/token_extensions/mod.rs
@@ -1,5 +1,0 @@
-mod token_extension_helpers;
-mod token_extensions_parser;
-
-pub use token_extension_helpers::*;
-pub use token_extensions_parser::*;
