--- crates/parser/src/account_parser/token_program/mod.rs
@@ -1,3 +1,0 @@
-mod token_program_parser;
-
-pub use token_program_parser::*;
