--- crates/parser/src/ix_parser/mod.rs
@@ -1,7 +1,0 @@
-#[cfg(feature = "token-program")]
-pub mod token_program;
-
-#[cfg(feature = "token-extensions")]
-pub mod token_extensions;
-
-pub mod helpers;
