--- crates/parser/src/ix_parser/token_extensions/mod.rs
@@ -1,4 +1,0 @@
-mod extensions;
-mod token_extension_parser;
-mod token_extensions_ix;
-pub use token_extension_parser::*;
