--- crates/parser/src/ix_parser/token_program/mod.rs
@@ -1,4 +1,0 @@
-mod token_program_parser;
-
-pub use token_program_parser::*;
-pub mod token_ix;
