+++ crates/parser/src/lib.rs
@@ -1,3 +1,7 @@
-pub mod account_parser;
+#[cfg(feature = "token-extensions")]
+pub mod token_extension_program;
 
-pub mod ix_parser;
+#[cfg(feature = "token-program")]
+pub mod token_program;
+
+mod helpers;
