+++ crates/parser/src/token_extension_program/ix_parser/extensions/mod.rs
@@ -1,6 +1,7 @@
 mod common_ix;
 mod confidential_transfer;
 mod confidential_transfer_fee;
+pub mod helpers;
 mod token_group;
 mod token_metadata;
 mod transfer_fee;
@@ -11,5 +12,3 @@ pub use confidential_transfer_fee::*;
 pub use token_group::*;
 pub use token_metadata::*;
 pub use transfer_fee::*;
-
-pub mod helpers;
