+++ crates/parser/src/token_extension_program/ix_parser/mod.rs
@@ -1,0 +1,9 @@
+mod parser;
+
+mod ixs;
+
+mod extensions;
+
+pub use extensions::*;
+pub use ixs::*;
+pub use parser::*;
