+++ crates/parser/src/token_program/ix_parser/parser.rs
@@ -4,8 +4,8 @@ use yellowstone_vixen_core::{
     ReadableInstruction,
 };
 
-use super::token_ix::*;
-use crate::ix_parser::helpers::{
+use super::ixs::*;
+use crate::helpers::{
     check_min_accounts_req, check_pubkeys_match, get_multisig_signers, to_supported_coption_pubkey,
     to_supported_pubkey,
 };
